requirements/base.txt
Django>2.0,<2.1
django-environ
rapidpro-python
